backend/render-requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
supabase>=2.0.0
requests-oauthlib>=2.0.0
cryptography>=42.0.8
python-dotenv>=1.0.1
pymongo==4.5.0
pydantic>=2.6.4
email-validator>=2.2.0
pyjwt>=2.10.1
passlib>=1.7.4
tzdata>=2024.2
motor==3.3.1
python-jose>=3.3.0
requests>=2.31.0
pandas>=2.2.0
numpy>=1.26.0
python-multipart>=0.0.9
typer>=0.9.0
httpx>=0.25.0
websockets>=12.0
bcrypt>=4.0.1
python-dateutil>=2.8.2
aiofiles>=23.1.0